requirements.txt
opencv-python
facedetector-py
filetype
numpy
pytest
tf-keras
scikit-learn
tqdm
